third_party/blink/web_tests/platform/mac-mac14-arm64/fast/webgl/texImage-imageBitmap-from-imageBitmap-from-offscreen-canvas-expected.txt

CONSOLE ERROR: Uncaught TypeError: Cannot read properties of null (reading 'clearColor')
FAIL Unexpected error: Uncaught TypeError: Cannot read properties of null (reading 'clearColor')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

Test FAILED
